feat(adoption): 다른 분양건 카드 피그마 반영 및 상세 컴포넌트 분리(SRP) (#137)
- 다른 분양건 카드 피그마 card-2 재구현(이미지 280x210, 배지, 소개 3줄, 통계+관심/공유)
- 섹션 제목 위 48px, gap 39px, 컬럼 920px 중앙(외곽서 260px)
- 첫 사진(대표사진)에서 왼쪽 화살표 숨김
- FavoriteShareActions 공통 추출, 배지 raw span → 공통 Badge
- AdoptionDetailContent SRP 분리: AdoptionDetailHero/HeroImageCarousel/AdoptionCtaBar
